requirements.txt
anytree
Flask
Flask-SQLAlchemy
Flask-Testing
Pillow
psycopg2-binary
PyGuitarPro
requests
SQLAlchemy
uwsgi
https://github.com/NonSvizzero/python-mingus/archive/master.zip
